requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
automat==25.4.16
autopep8==2.3.2
billiard==4.2.4
boolean-py==5.0
boto3==1.42.83
botocore==1.42.83
cachecontrol==0.14.4
cbor2==5.9.0
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels-redis==4.3.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==46.0.6
cyclonedx-python-lib==11.7.0
daphne==4.2.1
defusedxml==0.7.1
django==6.0.3
django-allauth==65.15.1
django-allauth-ui==1.8.1
django-axes==8.3.1
django-celery-beat==2.9.0
django-celery-results==2.6.0
django-cors-headers==4.9.0
django-dbbackup==5.2.0
django-debug-toolbar==6.3.0
django-phonenumber-field==8.4.0
django-redis==6.0.0
django-silk==5.5.0
django-smart-ratelimit==2.2.0
django-storages==1.14.6
django-taggit==6.1.0
django-timezone-field==7.2.1
django-tinymce==5.0.0
django-widget-tweaks==1.5.1
filelock==3.25.2
gprof2dot==2025.4.14
gunicorn==25.3.0
hiredis==3.3.1
hyperlink==21.0.0
idna==3.11
incremental==24.11.0
iniconfig==2.3.0
jmespath==1.1.0
kombu==5.6.2
license-expression==30.4.4
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
packageurl-python==0.17.6
packaging==26.0
phonenumbers==9.0.27
pillow==12.2.0
pip==26.0.1
pip-api==0.0.34
pip-audit==2.10.0
pip-requirements-parser==32.0.1
platformdirs==4.9.4
pluggy==1.6.0
prompt-toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
py-serializable==2.1.0
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pygments==2.20.0
pyjwt==2.12.1
pyopenssl==26.0.0
pyparsing==3.3.2
pytest==9.0.2
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.1.0
pyyaml==6.0.3
redis==7.4.0
requests==2.33.1
rich==14.3.3
s3transfer==0.16.0
sentry-sdk==2.57.0
service-identity==24.2.0
six==1.17.0
slippers==0.6.3
sortedcontainers==2.4.0
sqlparse==0.5.5
tomli==2.4.1
tomli-w==1.2.0
twisted==25.5.0
txaio==25.12.2
typeguard==2.13.3
typing-extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
ujson==5.12.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
zope-interface==8.2
